Rename `kernels_body.py` in `SolverFeatherstone`
